+++ NEWS.md
@@ -4,4 +4,9 @@
 
 ## Version: 0.0.4
 
-* Applying mortality to each generation separately rather than on selection candidates, which may come from more than a single generation. 
+* Applying mortality to each generation separately rather than on selection candidates, which may come from more than a single generation.
+
+## Version: 0.1.0
+
+* Added function `assortative` for assortative/disassortative mating.
+* Added `"-P"` and `"-PA"` to arguments `fsel` and `msel` to select in opposite directions of `"P"` (phenotype) and `"PA"` (parent average).